chatApp-API/requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
daphne==2.5.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
gunicorn==20.0.4
hiredis==1.1.0
html5lib==1.0.1
hyperlink==20.0.1
idna==2.8
importlib-metadata==4.6.1
incremental==17.5.0
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.0.1
lockfile==0.12.2
MarkupSafe==2.0.1
msgpack==1.0.0
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
zipp==3.5.0
zope.interface==5.1.2
